Improvements from second test print
On the lower section, cut out a notch to fit the power connectors on
the replicape.  Further clean out the ethernet/USB notch and widen
it, since I had to cut out a little bit to fit the BBG into this print.
Increase the size of the USB cutout slightly to be more likely to fit
both the BBB mini-USB and BBG micro-USB in the same location.  Make
the bottom flange extend with holes on all four sides because the
best mounting orientation is unclear.  Add thickness to flange around
mounting holes to accomodate mounting screws.

On the upper section, make another attempt to get the screw hole locations
correct for the 80mm fan.  Reduce size of center face; probably don't need it
at all, but leaving it there for no good reason.